Refactor fixture import paths in test files
This commit mainly addresses the change in import paths for database connection objects and cursors across multiple test files. `tests.middleware.fixtures` has been moved to `tests.fixtures`, resulting in changes to import statements in the affected test files. This provides a cleaner organization and management of test cases, making code more maintainable.
