+++ tests/middleware/test_data_source_queries.py
@@ -12,7 +12,7 @@ from middleware.data_source_queries import (
 from tests.middleware.helper_functions import (
     get_boolean_dictionary,
 )
-from tests.middleware.fixtures import connection_with_test_data, dev_db_connection
+from tests.fixtures import connection_with_test_data, dev_db_connection
 
 
 @pytest.fixture
